refactor(api): reorganize routing structure and tests into modular files
- Split monolithic api.go into feature-based files (blocks, files, gateway, ipfs, pins, upload)
- Extract test helpers into test_helpers_test.go for better test organization
- Move gateway middleware to dedicated middleware/gateway.go
- Add gateway IPFS routing endpoint with dedicated handler
- Update API configuration and protocol interfaces
- Reorganize test structure with separate test files per endpoint
- Update go.mod with required dependencies for new structure
