+++ internal/api/api_test.go
@@ -1,44 +1,14 @@
 package api
 
 import (
-	"archive/zip"
-	"bytes"
-	"context"
-	"encoding/json"
-	"fmt"
-	"mime/multipart"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"strings"
 	"testing"
-	"time"
 
-	"github.com/google/uuid"
-	"github.com/ipfs/boxo/ipld/merkledag"
-	blocks "github.com/ipfs/go-block-format"
-	"github.com/ipfs/go-cid"
-	"github.com/ipld/go-car/v2"
-	"github.com/ipld/go-car/v2/blockstore"
-	"github.com/multiformats/go-multicodec"
-	"github.com/multiformats/go-multihash"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	pluginCore "go.lumeweb.com/portal-plugin-ipfs/core"
 	"go.lumeweb.com/portal-plugin-ipfs/internal"
-	"go.lumeweb.com/portal-plugin-ipfs/internal/api/dto"
-	"go.lumeweb.com/portal-plugin-ipfs/internal/config"
-	pluginDb "go.lumeweb.com/portal-plugin-ipfs/internal/db"
+	pluginConfig "go.lumeweb.com/portal-plugin-ipfs/internal/config"
 	"go.lumeweb.com/portal-plugin-ipfs/internal/testing/mocks"
 	"go.lumeweb.com/portal-plugin-ipfs/internal/testing/util"
-	"go.lumeweb.com/portal/core"
 	coreTesting "go.lumeweb.com/portal/core/testing"
-	coreMocks "go.lumeweb.com/portal/core/testing/mocks"
-	"go.lumeweb.com/portal/db/models"
-	"go.lumeweb.com/portal/db/types"
-	"go.lumeweb.com/queryutil"
-	"go.lumeweb.com/queryutil/filter"
-	"gorm.io/gorm"
 )
 
 func TestMain(m *testing.M) {
@@ -47,1070 +17,27 @@ func TestMain(m *testing.M) {
 		coreTesting.WithMockServiceFactory(pluginCore.PIN_SERVICE, mocks.NewMockIPFSPinService),
 		coreTesting.WithMockServiceFactory(pluginCore.BLOCK_SERVICE, mocks.NewMockBlockService),
 		coreTesting.WithMockServiceFactory(pluginCore.UPLOAD_SERVICE, mocks.NewMockUploadService),
-		coreTesting.WithHTTPService(),
-		coreTesting.WithPlugins(),
+		coreTesting.WithMockServiceFactory(pluginCore.WEBSITE_SERVICE, mocks.NewMockWebsiteService),
 		coreTesting.WithAPI(internal.ProtocolName, NewAPI),
 		coreTesting.WithAPIID(internal.ProtocolName),
 
 		util.GetProtocolMock(),
-		coreTesting.WithProtocolConfig(internal.ProtocolName, config.ProtocolConfig{}),
+		coreTesting.WithProtocolConfig(internal.ProtocolName, pluginConfig.ProtocolConfig{}),
 	)
 }
 
-// mockHelper provides common mock setup functions for tests
-type mockHelper struct {
-	ctx coreTesting.TestContext
-	t   *testing.T
-}
-
-func newMockHelper(t *testing.T, ctx coreTesting.TestContext) *mockHelper {
-	return &mockHelper{
-		ctx: ctx,
-		t:   t,
-	}
-}
-
-// createMockIPFSPin creates a standardized IPFSPin mock object
-func createMockIPFSPin(userID uint, testCID cid.Cid, pinID types.BinaryUUID, status pluginDb.PinningStatus) *pluginDb.IPFSPin {
-	return &pluginDb.IPFSPin{
-		RequestID: pinID,
-		UserID:    userID,
-		CID:       testCID.Bytes(),
-		Name:      "test",
-		Status:    status,
-		Origins:   nil,
-		Meta:      nil,
-		Delegates: nil,
-		Info:      nil,
-	}
-}
-
-// createMockUnixFSNode creates a standardized UnixFSNode mock object
-func createMockUnixFSNode(testCID cid.Cid) *pluginDb.UnixFSNode {
-	return &pluginDb.UnixFSNode{
-		Name:      "test",
-		Type:      2,
-		BlockSize: 1024,
-		ChildCID:  []cid.Cid{testCID},
-	}
-}
-
-// SetupPinServiceMocks configures common pin service mock expectations
-func (m *mockHelper) SetupPinServiceMocks(userID uint, testCID cid.Cid, pinID types.BinaryUUID) *mocks.MockIPFSPinService {
-	mockPinService := core.GetService[*mocks.MockIPFSPinService](m.ctx, pluginCore.PIN_SERVICE)
-
-	// Setup AddPin expectation
-	mockPinService.EXPECT().AddPin(mock.Anything, mock.AnythingOfType("*db.IPFSPin")).Return(createMockIPFSPin(userID, testCID, pinID, pluginDb.PinningStatusQueued), nil).Maybe()
-
-	// Setup GetPinByRequestID expectation
-	mockPinService.EXPECT().GetPinByRequestID(mock.Anything, pinID).Return(createMockIPFSPin(userID, testCID, pinID, pluginDb.PinningStatusPinned), nil).Maybe()
-
-	// Setup ListPins expectation
-	mockPinService.EXPECT().ListPins(mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.IPFSPin{createMockIPFSPin(userID, testCID, pinID, pluginDb.PinningStatusPinned)}, int64(1), nil).Maybe()
-
-	// Setup DeletePin expectation
-	mockPinService.EXPECT().DeletePin(mock.Anything, pinID).Return(nil).Maybe()
-
-	// Setup ReplacePin expectation
-	mockPinService.EXPECT().ReplacePin(mock.Anything, mock.AnythingOfType("uint"), mock.AnythingOfType("string"), pinID, mock.AnythingOfType("*db.IPFSPin")).Return(createMockIPFSPin(userID, testCID, pinID, pluginDb.PinningStatusPinned), nil).Maybe()
-
-	// Setup GetPinByCIDAndUser expectation
-	mockPinService.EXPECT().GetPinByCIDAndUser(mock.Anything, mock.AnythingOfType("cid.Cid"), userID).Return(createMockIPFSPin(userID, testCID, pinID, pluginDb.PinningStatusPinned), nil).Maybe()
-
-	return mockPinService
-}
-
-// SetupBlockServiceMocks configures common block service mock expectations
-func (m *mockHelper) SetupBlockServiceMocks(testCID cid.Cid) *mocks.MockBlockService {
-	mockBlockService := core.GetService[*mocks.MockBlockService](m.ctx, pluginCore.BLOCK_SERVICE)
-
-	mockBlockService.EXPECT().GetBlockMeta(mock.Anything, testCID).Return(createMockUnixFSNode(testCID), nil).Maybe()
-
-	mockBlockService.EXPECT().GetBlockMetaBatch(mock.Anything, mock.AnythingOfType("[]cid.Cid")).Return(map[string]*pluginDb.UnixFSNode{
-		testCID.String(): createMockUnixFSNode(testCID),
-	}, nil).Maybe()
-
-	return mockBlockService
-}
-
-// SetupUploadServiceMocks configures common upload service mock expectations
-func (m *mockHelper) SetupUploadServiceMocks(expectedCID cid.Cid) *mocks.MockUploadService {
-	mockUploadService := core.GetService[*mocks.MockUploadService](m.ctx, pluginCore.UPLOAD_SERVICE)
-
-	mockUploadService.EXPECT().HandleUpload(mock.Anything, mock.Anything, mock.AnythingOfType("uint")).Return(expectedCID, "", nil).Maybe()
-
-	return mockUploadService
-}
-
-// SetupFileManagerServiceMocks configures file manager service
-func (m *mockHelper) SetupFileManagerServiceMocks() *mocks.MockFileManagerService {
-	mockFileManagerService := core.GetService[*mocks.MockFileManagerService](m.ctx, pluginCore.FILE_MANAGER_SERVICE)
-
-	return mockFileManagerService
-}
-
-// SetupWorkflowServiceMock configures workflow service mock for tests that need pin workflows
-func (m *mockHelper) SetupWorkflowServiceMock() {
-	// Setup core.WorkflowService mock using the higher-level helper
-	mockWorkflowService := core.GetService[*coreTesting.MockWorkflowService](m.ctx, core.WORKFLOW_SERVICE)
-	// StartWorkflow is called with: context, workflowName, and 5 workflow options = 6 total workflow options
-	mockWorkflowService.ExpectStartWorkflowWithExactArgs("ipfs.network.pin", (*models.Request)(nil), nil, 5)
-}
-
-// SetupUploadWorkflowServiceMock configures workflow service mock for tests that need upload workflows
-func (m *mockHelper) SetupUploadWorkflowServiceMock() {
-	// Setup core.WorkflowService mock using the higher-level helper
-	mockWorkflowService := core.GetService[*coreTesting.MockWorkflowService](m.ctx, core.WORKFLOW_SERVICE)
-	// StartWorkflow is called with: context, workflowName, and 5 workflow options = 6 total workflow options
-	mockWorkflowService.ExpectStartWorkflowWithExactArgs("ipfs.upload", (*models.Request)(nil), nil, 5)
-}
-
-// SetupAllCommonMocks configures all common service mocks for basic test scenarios
-func (m *mockHelper) SetupAllCommonMocks(userID uint, testCID cid.Cid, pinID types.BinaryUUID) {
-	m.SetupPinServiceMocks(userID, testCID, pinID)
-	m.SetupBlockServiceMocks(testCID)
-	m.SetupFileManagerServiceMocks()
-}
-
-// SetupAuthenticatedTest creates a test user, logs them in, and sets up common mocks
-func (m *mockHelper) SetupAuthenticatedTest() (string, uint, cid.Cid, types.BinaryUUID) {
-	token, userID := createTestUserAndLogin(m.ctx)
-	testCID := cid.MustParse("bafybeieffnocaq7t4w4daagvydl32igft5oziyyaebqr6vx6rb3fwh2ab4")
-	pinID := types.NewBinUUID()
-
-	m.SetupAllCommonMocks(userID, testCID, pinID)
-
-	return token, userID, testCID, pinID
-}
-
-// SetupAuthenticatedTestWithCID creates a test user, logs them in, and sets up mocks with custom CID
-func (m *mockHelper) SetupAuthenticatedTestWithCID(testCID cid.Cid) (string, uint) {
-	token, userID := createTestUserAndLogin(m.ctx)
-	pinID := types.NewBinUUID()
-
-	m.SetupAllCommonMocks(userID, testCID, pinID)
-
-	return token, userID
-}
-
-func createTestUserAndLogin(ctx coreTesting.TestContext) (string, uint) {
-	mockAuth := core.GetService[*coreTesting.MockAuthService](ctx, core.AUTH_SERVICE)
-	token, user, err := mockAuth.CreateAndLoginUser("test@example.com", "example")
-	if err != nil {
-		ctx.T().Fatalf("failed to create and login test user: %v", err)
-	}
-
-	return token, user.ID
-}
-
-func setAuthHeader(req *http.Request, token string) {
-	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
-}
-
-// makeAuthenticatedRequest creates and executes an authenticated API request, returning the response
-func (m *mockHelper) makeAuthenticatedRequest(method, url string, token string, body []byte) *httptest.ResponseRecorder {
-	req := m.ctx.NewAPIRequest(method, url, body)
-	setAuthHeader(req, token)
-	rec := httptest.NewRecorder()
-	m.ctx.Router().ServeHTTP(rec, req)
-	return rec
-}
-
-// assertJSONResponse is a helper to assert JSON response structure
-func (m *mockHelper) assertJSONResponse(t *testing.T, rec *httptest.ResponseRecorder, expectedStatus int, target interface{}) {
-	assert.Equal(t, expectedStatus, rec.Code)
-	err := json.Unmarshal(rec.Body.Bytes(), target)
-	assert.NoError(t, err)
-}
-
-// createTestFileDataSet creates a standardized set of test file paths for common scenarios
-func createTestFileDataSet(userID uint, testCID1, testCID2, testCID3 cid.Cid) []*pluginDb.FilePath {
-	return []*pluginDb.FilePath{
-		createMockFilePath(userID, testCID1, "/file1.txt", "file1.txt", false),
-		createMockFilePath(userID, testCID2, "/file2.txt", "file2.txt", false),
-		createMockFilePath(userID, testCID3, "/file3.txt", "file3.txt", false),
-	}
-}
-
-// createTestDirectoryDataSet creates a standardized set of test file paths with directories
-func createTestDirectoryDataSet(userID uint, testCID1, testCID2, testCID3 cid.Cid) []*pluginDb.FilePath {
-	return []*pluginDb.FilePath{
-		createMockFilePath(userID, testCID1, "/file1.txt", "file1.txt", false),
-		createMockFilePath(userID, testCID2, "/test_dir", "test_dir", true),
-		createMockFilePath(userID, testCID3, "/test_dir/file2.txt", "file2.txt", false),
-	}
-}
-
-// createBreadcrumbsDataSet creates a standardized set of test file paths for breadcrumb tests
-func createBreadcrumbsDataSet(userID uint, testCID1, testCID2, testCID3 cid.Cid) []*pluginDb.FilePath {
-	return []*pluginDb.FilePath{
-		createMockFilePath(userID, testCID1, "/test_dir", "test_dir", true),
-		createMockFilePath(userID, testCID2, "/test_dir/subdir", "subdir", true),
-		createMockFilePath(userID, testCID3, "/test_dir/subdir/file.txt", "file.txt", false),
-	}
-}
-
-// setupMultiplePinServiceMocks configures pin service mocks for multiple CIDs
-func (m *mockHelper) setupMultiplePinServiceMocks(userID uint, testCIDs []cid.Cid) {
-	mockPinService := core.GetService[*mocks.MockIPFSPinService](m.ctx, pluginCore.PIN_SERVICE)
-
-	for _, testCID := range testCIDs {
-		pinID := types.NewBinUUID()
-		mockPinService.EXPECT().GetPinByCIDAndUser(mock.Anything, mock.AnythingOfType("cid.Cid"), userID).Return(
-			createMockIPFSPin(userID, testCID, pinID, pluginDb.PinningStatusPinned), nil).Maybe()
-	}
-}
-
-func createMockFilePath(userID uint, testCID cid.Cid, path, name string, isDirectory bool) *pluginDb.FilePath {
-	parentPath := util.CalculateParentPath(path)
-
-	depth := 0
-	if path != "/" {
-		segments := strings.Split(strings.Trim(path, "/"), "/")
-		depth = len(segments)
-	}
-
-	return &pluginDb.FilePath{
-		UserID:      userID,
-		CID:         testCID.Bytes(),
-		Path:        path,
-		Name:        name,
-		Type:        0,
-		Size:        1024,
-		IsDirectory: isDirectory,
-		IsOrphan:    false,
-		ParentPath:  parentPath,
-		Depth:       depth,
-	}
-}
-
-type pinTestHelper struct {
-	token string
-	pinID string
-	cid   cid.Cid
-}
-
-func setupPinTest(t *testing.T, ctx coreTesting.TestContext) *pinTestHelper {
-	helper := newMockHelper(t, ctx)
-	token, _, testCID, pinID := helper.SetupAuthenticatedTest()
-
-	return &pinTestHelper{
-		token: token,
-		pinID: pinID.String(),
-		cid:   testCID,
-	}
-}
-
-func createTestCAR(t *testing.T) ([]byte, cid.Cid) {
-	content := []byte("test file content")
-
-	pref := cid.Prefix{
-		Version:  1,
-		Codec:    uint64(multicodec.Raw),
-		MhType:   multihash.SHA2_256,
-		MhLength: -1,
-	}
-
-	c, err := pref.Sum(content)
-	if err != nil {
-		t.Fatalf("failed to create CID: %v", err)
-	}
-
-	tmpFile, err := os.CreateTemp("", "test-car-*.car")
-	if err != nil {
-		t.Fatalf("failed to create temp file: %v", err)
-	}
-	defer os.Remove(tmpFile.Name())
-	defer tmpFile.Close()
-
-	cdest, err := blockstore.OpenReadWrite(tmpFile.Name(), []cid.Cid{c}, car.WriteAsCarV1(true))
-	if err != nil {
-		t.Fatalf("failed to create CAR writer: %v", err)
-	}
-
-	blk, err := blocks.NewBlockWithCid(content, c)
-	if err != nil {
-		t.Fatalf("failed to create block: %v", err)
-	}
-	err = cdest.Put(context.Background(), blk)
-	if err != nil {
-		t.Fatalf("failed to write block: %v", err)
-	}
-
-	if err := cdest.Finalize(); err != nil {
-		t.Fatalf("failed to finalize CAR: %v", err)
-	}
-
-	carData, err := os.ReadFile(tmpFile.Name())
-	if err != nil {
-		t.Fatalf("failed to read temp CAR file: %v", err)
-	}
-
-	return carData, c
-}
-
-func TestAPI_listPins(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _, _, _ := helper.SetupAuthenticatedTest()
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, "/api/pins", token, nil)
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-		var response dto.PinResultsResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-		assert.NotEmpty(t, response.Results)
-	})
-}
-
-func TestAPI_addPin(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		helper.SetupWorkflowServiceMock()
-		token, _, _, _ := helper.SetupAuthenticatedTest()
-
-		reqBody := `{"cid":"bafybeieffnocaq7t4w4daagvydl32igft5oziyyaebqr6vx6rb3fwh2ab4","name":"test"}`
-		rec := helper.makeAuthenticatedRequest(http.MethodPost, "/api/pins", token, []byte(reqBody))
-
-		assert.Equal(t, http.StatusAccepted, rec.Code)
-		var response dto.PinStatusResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-		assert.NotPanics(t, func() {
-			uuid.MustParse(response.RequestID)
-		})
-	})
-}
-
-func TestAPI_getPin(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _, _, pinID := helper.SetupAuthenticatedTest()
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, fmt.Sprintf("/api/pins/%s", pinID.String()), token, nil)
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-		var response dto.PinStatusResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-		assert.NotPanics(t, func() {
-			uuid.MustParse(response.RequestID)
-		})
-	})
-}
-
-func TestAPI_replacePin(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		helper.SetupWorkflowServiceMock()
-		token, _, _, pinID := helper.SetupAuthenticatedTest()
-
-		reqBody := `{"cid":"bafybeieffnocaq7t4w4daagvydl32igft5oziyyaebqr6vx6rb3fwh2ab4","name":"test"}`
-		rec := helper.makeAuthenticatedRequest(http.MethodPost, fmt.Sprintf("/api/pins/%s", pinID.String()), token, []byte(reqBody))
-
-		assert.Equal(t, http.StatusAccepted, rec.Code)
-		var response dto.PinStatusResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-		assert.NotPanics(t, func() {
-			uuid.MustParse(response.RequestID)
-		})
-	})
-}
-
-func TestAPI_deletePin(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _, _, pinID := helper.SetupAuthenticatedTest()
-
-		rec := helper.makeAuthenticatedRequest(http.MethodDelete, fmt.Sprintf("/api/pins/%s", pinID.String()), token, nil)
-
-		assert.Equal(t, http.StatusAccepted, rec.Code)
-	})
-}
-
-func TestAPI_listFiles(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		testCID1 := util.GenerateTestCID(t, "test data 1")
-		testCID2 := util.GenerateTestCID(t, "test data 2")
-
-		// Create mock file paths using helper
-		mockFilePaths := []*pluginDb.FilePath{
-			createMockFilePath(userID, testCID1, "/file1.txt", "file1.txt", false),
-			createMockFilePath(userID, testCID2, "/file2.txt", "file2.txt", false),
-		}
-
-		// Setup file manager service mock to return our test data
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return(mockFilePaths, int64(2), nil)
-
-		// Setup pin service mock for multiple CIDs
-		helper.setupMultiplePinServiceMocks(userID, []cid.Cid{testCID1, testCID2})
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, "/api/files", token, nil)
-
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-
-		assert.Equal(t, int64(2), response.Total)
-		assert.Len(t, response.Data, 2)
-
-		for _, item := range response.Data {
-			assert.NotEmpty(t, item.Path)
-			assert.NotEmpty(t, item.Name)
-			assert.IsType(t, uint64(0), item.Size)
-			assert.IsType(t, false, item.IsDirectory)
-			assert.IsType(t, 0, item.Depth)
-			assert.IsType(t, time.Time{}, item.Created)
-			assert.IsType(t, time.Time{}, item.Updated)
-		}
-	})
-}
-
-func TestAPI_listFiles_Unauthorized(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		req := ctx.NewAPIRequest(http.MethodGet, "/api/files", nil)
-		rec := httptest.NewRecorder()
-		ctx.Router().ServeHTTP(rec, req)
-		assert.Equal(t, http.StatusUnauthorized, rec.Code)
-	})
-}
-
-func TestAPI_listFiles_EmptyResults(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		// Setup file manager service mock to return empty results
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{}, int64(0), nil)
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, "/api/files", token, nil)
-
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(0), response.Total)
-		assert.Empty(t, response.Data)
-	})
-}
-
-func TestAPI_listFiles_WithFilters(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		testCID1 := util.GenerateTestCID(t, "test data 1")
-		testCID2 := util.GenerateTestCID(t, "test data 2")
-		testCID3 := util.GenerateTestCID(t, "test data 3")
-
-		// Create mock file paths using helper
-		mockFilePaths := createTestDirectoryDataSet(userID, testCID1, testCID2, testCID3)
-
-		// Setup file manager service mock to return filtered results for each request
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-
-		// Setup pin service mock for multiple CIDs
-		helper.setupMultiplePinServiceMocks(userID, []cid.Cid{testCID1, testCID2, testCID3})
-
-		// Setup expectations for each filtered request
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[0]}, int64(1), nil).Times(1)
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[1]}, int64(1), nil).Times(1)
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[0], mockFilePaths[2]}, int64(2), nil).Times(1)
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[0], mockFilePaths[1]}, int64(2), nil).Times(1)
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[0], mockFilePaths[2]}, int64(2), nil).Times(1)
-
-		// Test filter by name = file1.txt
-		url, err := queryutil.BuildURL("/api/files", nil, nil, filter.Equal("name", "file1.txt"))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(1), response.Total)
-		assert.Len(t, response.Data, 1)
-		assert.Equal(t, "file1.txt", response.Data[0].Name)
-
-		// Test filter by is_directory = true
-		url, err = queryutil.BuildURL("/api/files", nil, nil, filter.Equal("is_directory", true))
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(1), response.Total)
-		assert.Len(t, response.Data, 1)
-		assert.True(t, response.Data[0].IsDirectory)
-		assert.Equal(t, "test_dir", response.Data[0].Name)
-
-		// Test filter by name contains "file"
-		url, err = queryutil.BuildURL("/api/files", nil, nil, filter.Contains("name", "file"))
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(2), response.Total)
-		assert.Len(t, response.Data, 2)
-
-		// Test OR filter
-		url, err = queryutil.BuildURL("/api/files", nil, nil, filter.Or(
-			filter.Equal("name", "file1.txt"),
-			filter.Equal("is_directory", true),
-		))
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(2), response.Total)
-		assert.Len(t, response.Data, 2)
-
-		// Test AND filter
-		url, err = queryutil.BuildURL("/api/files", nil, nil, filter.And(
-			filter.Contains("name", "file"),
-			filter.Equal("is_directory", false),
-		))
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(2), response.Total)
-		assert.Len(t, response.Data, 2)
-		for _, result := range response.Data {
-			assert.Contains(t, result.Name, "file")
-			assert.False(t, result.IsDirectory)
-		}
-	})
-}
-
-func TestAPI_listFiles_Pagination(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		testCID1 := util.GenerateTestCID(t, "test data 1")
-		testCID2 := util.GenerateTestCID(t, "test data 2")
-		testCID3 := util.GenerateTestCID(t, "test data 3")
-
-		// Create mock file paths using helper
-		mockFilePaths := createTestFileDataSet(userID, testCID1, testCID2, testCID3)
-
-		// Setup file manager service mock to return paginated results for each request
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-
-		// Setup pin service mock for multiple CIDs
-		helper.setupMultiplePinServiceMocks(userID, []cid.Cid{testCID1, testCID2, testCID3})
-
-		// Setup expectations for each paginated request
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[0], mockFilePaths[1]}, int64(3), nil).Times(1)
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[1], mockFilePaths[2]}, int64(3), nil).Times(1)
-		mockFileManagerService.EXPECT().ListFiles(mock.Anything, userID, mock.Anything, mock.Anything, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[0]}, int64(3), nil).Times(1)
-
-		// Test first page: _start=0&_end=2
-		pagination, err := filter.NewPagination(0, 2)
-		assert.NoError(t, err)
-		url, err := queryutil.BuildURL("/api/files", nil, &pagination)
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(3), response.Total)
-		assert.Len(t, response.Data, 2)
-
-		// Test second page: _start=1&_end=3
-		pagination, err = filter.NewPagination(1, 2)
-		assert.NoError(t, err)
-		url, err = queryutil.BuildURL("/api/files", nil, &pagination)
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(3), response.Total)
-		assert.Len(t, response.Data, 2)
-
-		// Test single item: _start=0&_end=1
-		pagination, err = filter.NewPagination(0, 1)
-		assert.NoError(t, err)
-		url, err = queryutil.BuildURL("/api/files", nil, &pagination)
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(3), response.Total)
-		assert.Len(t, response.Data, 1)
-	})
-}
-
-func TestAPI_listDirectoryContents(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		testCID1 := util.GenerateTestCID(t, "test data 1")
-		testCID2 := util.GenerateTestCID(t, "test data 2")
-		testCID3 := util.GenerateTestCID(t, "test data 3")
-
-		// Create mock file paths using helper
-		mockFilePaths := createTestDirectoryDataSet(userID, testCID1, testCID2, testCID3)
-
-		// Setup file manager service mock to return only root directory contents
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().ListDirectoryContents(mock.Anything, userID, mock.Anything).Return([]*pluginDb.FilePath{mockFilePaths[1], mockFilePaths[0]}, nil)
-
-		url, err := queryutil.BuildURL("/api/files/directory", nil, nil, filter.Equal("parent_path", "/"))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-
-		assert.Equal(t, int64(2), response.Total)
-		assert.Len(t, response.Data, 2)
-
-		assert.True(t, response.Data[0].IsDirectory)
-		assert.False(t, response.Data[1].IsDirectory)
-		assert.Equal(t, "test_dir", response.Data[0].Name)
-		assert.Equal(t, "file1.txt", response.Data[1].Name)
-	})
-}
-
-func TestAPI_listDirectoryContents_Unauthorized(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		url, err := queryutil.BuildURL("/api/files/directory", nil, nil, filter.Equal("parent_path", "/"))
-		assert.NoError(t, err)
-		req := ctx.NewAPIRequest(http.MethodGet, url, nil)
-		rec := httptest.NewRecorder()
-		ctx.Router().ServeHTTP(rec, req)
-		assert.Equal(t, http.StatusUnauthorized, rec.Code)
-	})
-}
-
-func TestAPI_listDirectoryContents_EmptyDirectory(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		// Setup file manager service mock to return empty results
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().ListDirectoryContents(mock.Anything, userID, mock.Anything).Return([]*pluginDb.FilePath{}, nil)
-
-		url, err := queryutil.BuildURL("/api/files/directory", nil, nil, filter.Equal("parent_path", "/empty"))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(0), response.Total)
-		assert.Empty(t, response.Data)
-	})
-}
-
-func TestAPI_listDirectoryContents_NonExistentUser(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		// Setup file manager service mock to return empty results
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().ListDirectoryContents(mock.Anything, userID, mock.Anything).Return([]*pluginDb.FilePath{}, nil)
-
-		url, err := queryutil.BuildURL("/api/files/directory", nil, nil, filter.Equal("parent_path", "/"))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-		assert.Equal(t, int64(0), response.Total)
-	})
-}
-
-func TestAPI_getBreadcrumbs(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		testCID1 := util.GenerateTestCID(t, "test data 1")
-		testCID2 := util.GenerateTestCID(t, "test data 2")
-		testCID3 := util.GenerateTestCID(t, "test data 3")
-
-		// Create mock file paths using helper
-		mockFilePaths := createBreadcrumbsDataSet(userID, testCID1, testCID2, testCID3)
-
-		// Setup file manager service mock to return our test data
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().GetBreadcrumbs(mock.Anything, userID, mock.Anything).Return(mockFilePaths, nil)
-
-		url, err := queryutil.BuildURL("/api/files/breadcrumbs", nil, nil, filter.Equal("path", "/test_dir/subdir/file.txt"))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-
-		var response queryutil.Response[[]dto.FileManagerItem]
-		helper.assertJSONResponse(t, rec, http.StatusOK, &response)
-
-		assert.Equal(t, int64(3), response.Total)
-		assert.Len(t, response.Data, 3)
-
-		assert.Equal(t, "/test_dir", response.Data[0].Path)
-		assert.Equal(t, "/test_dir/subdir", response.Data[1].Path)
-		assert.Equal(t, "/test_dir/subdir/file.txt", response.Data[2].Path)
-	})
-}
-
-func TestAPI_getBreadcrumbs_Unauthorized(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		url, err := queryutil.BuildURL("/api/files/breadcrumbs", nil, nil, filter.Equal("path", "/test"))
-		assert.NoError(t, err)
-		req := ctx.NewAPIRequest(http.MethodGet, url, nil)
-		rec := httptest.NewRecorder()
-		ctx.Router().ServeHTTP(rec, req)
-		assert.Equal(t, http.StatusUnauthorized, rec.Code)
-	})
-}
-
-func TestAPI_getBreadcrumbs_InvalidPath(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _ := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		url, err := queryutil.BuildURL("/api/files/breadcrumbs", nil, nil, filter.Equal("path", ""))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		assert.Equal(t, http.StatusBadRequest, rec.Code)
-
-		url, err = queryutil.BuildURL("/api/files/breadcrumbs", nil, nil, filter.Equal("path", "test/file.txt"))
-		assert.NoError(t, err)
-		rec = helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		assert.Equal(t, http.StatusBadRequest, rec.Code)
-	})
-}
-
-func TestAPI_getBreadcrumbs_PathNotFound(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, userID := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		// Setup file manager service mock to return error for non-existent path
-		mockFileManagerService := helper.SetupFileManagerServiceMocks()
-		mockFileManagerService.EXPECT().GetBreadcrumbs(mock.Anything, userID, "/nonexistent/file.txt").Return([]*pluginDb.FilePath{}, gorm.ErrRecordNotFound)
-
-		url, err := queryutil.BuildURL("/api/files/breadcrumbs", nil, nil, filter.Equal("path", "/nonexistent/file.txt"))
-		assert.NoError(t, err)
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, url, token, nil)
-		assert.Equal(t, http.StatusNotFound, rec.Code)
-	})
-}
-
-func TestAPI_handleUpload(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _ := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-		carData, expectedCID := createTestCAR(t)
-		helper.SetupUploadServiceMocks(expectedCID)
-		helper.SetupUploadWorkflowServiceMock()
-
-		body := &bytes.Buffer{}
-		writer := multipart.NewWriter(body)
-		part, _ := writer.CreateFormFile("file", "test.car")
-		part.Write(carData)
-		writer.Close()
-
-		req := ctx.NewAPIRequest(http.MethodPost, "/api/upload", body.Bytes())
-		setAuthHeader(req, token)
-		req.Header.Set("Content-Type", writer.FormDataContentType())
-		rec := httptest.NewRecorder()
-		ctx.Router().ServeHTTP(rec, req)
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-
-		var resp dto.PostUploadResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &resp)
-		assert.NoError(t, err)
-		assert.Equal(t, expectedCID.String(), resp.CID)
-	})
-}
-
-func TestAPI_handleGetBlockMeta(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _, testCID, _ := helper.SetupAuthenticatedTest()
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, fmt.Sprintf("/api/block/meta/%s", testCID.String()), token, nil)
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-		var response dto.BlockMetaResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-
-		assert.NotNil(t, response)
-		assert.IsType(t, "", response.Name)
-		assert.IsType(t, uint8(0), response.Type)
-		assert.IsType(t, int64(0), response.BlockSize)
-		assert.IsType(t, []string{}, response.ChildCID)
-		assert.True(t, len(response.ChildCID) > 0, "ChildCID should not be empty")
-	})
-}
-
-func TestAPI_handleGetBlockMetaBatch(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _, testCID, _ := helper.SetupAuthenticatedTest()
-
-		reqBody := fmt.Sprintf(`{"cid":["%s"]}`, testCID.String())
-		rec := helper.makeAuthenticatedRequest(http.MethodPost, "/api/block/meta/batch", token, []byte(reqBody))
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-		var response map[string]*dto.BlockMetaResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-
-		assert.NotEmpty(t, response)
-		for cid, meta := range response {
-			assert.NotEmpty(t, cid)
-			assert.NotNil(t, meta)
-			assert.IsType(t, "", meta.Name)
-			assert.IsType(t, uint8(0), meta.Type)
-			assert.IsType(t, int64(0), meta.BlockSize)
-			assert.IsType(t, []string{}, meta.ChildCID)
-			assert.True(t, len(meta.ChildCID) > 0, "ChildCID should not be empty")
-		}
-	})
-}
-
-func TestAPI_handleGetInfo(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _ := helper.SetupAuthenticatedTestWithCID(util.GenerateTestCID(t, "test data"))
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, "/api/info", token, nil)
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-		var response dto.InfoResponse
-		err := json.Unmarshal(rec.Body.Bytes(), &response)
-		assert.NoError(t, err)
-
-		assert.NotEmpty(t, response.PeerID)
-		assert.NotEmpty(t, response.AnnouncementAddresses)
-		assert.NotEmpty(t, response.ConnectionAddresses)
-
-		for _, addr := range response.AnnouncementAddresses {
-			assert.Contains(t, addr, "/ip6/")
-			assert.Contains(t, addr, "/tcp/4001")
-		}
-
-		for _, addr := range response.ConnectionAddresses {
-			assert.Contains(t, addr, "/ip6/")
-			assert.Contains(t, addr, "/tcp/4001/p2p/")
-			assert.Contains(t, addr, response.PeerID)
-		}
-	})
-}
-
-func TestAPI_handleIPFSGet(t *testing.T) {
-	coreTesting.RunTestCase(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		helper := newMockHelper(t, ctx)
-		token, _, testCID, _ := helper.SetupAuthenticatedTest()
-
-		// Setup IPFS node mock expectations for HasBlock
-		protoMock := core.GetProtocol(internal.ProtocolName).(*mocks.MockProtoNode)
-		mockIPFSNode := protoMock.GetNode().(*mocks.MockIPFSNode)
-
-		// Mock HasBlock to return true for the test CID
-		mockIPFSNode.EXPECT().HasBlock(mock.Anything, testCID).Return(true, nil)
-
-		uploadSvc := core.GetService[*coreMocks.MockUploadService](ctx, core.UPLOAD_SERVICE)
-		testUpload := &models.Upload{
-			Model:    gorm.Model{ID: 1},
-			UserID:   1,
-			Hash:     testCID.Hash(),
-			CIDType:  1, // CIDv1
-			MimeType: "application/octet-stream",
-			Protocol: "ipfs",
-			Size:     1024,
-		}
-		uploadSvc.EXPECT().GetUpload(mock.Anything, internal.NewIPFSHash(testCID)).Return(testUpload, nil)
-
-		// Mock GetBlock to return a mock node for the test CID
-		testData := []byte("tornadocash")
-		mockNode := merkledag.NewRawNode(testData)
-		mockIPFSNode.EXPECT().GetBlock(mock.Anything, testCID).Return(mockNode, nil)
-
-		rec := helper.makeAuthenticatedRequest(http.MethodGet, fmt.Sprintf("/ipfs/%s", testCID.String()), token, nil)
-
-		assert.Equal(t, http.StatusOK, rec.Code)
-		assert.Contains(t, rec.Body.String(), "tornadocash")
-	})
-}
-
-// createTestZIPFile creates a test ZIP file for API testing
-func createTestZIPFile(t *testing.T, content string) *bytes.Buffer {
-	var buf bytes.Buffer
-	zipWriter := zip.NewWriter(&buf)
-	
-	writer, err := zipWriter.Create("test.txt")
-	require.NoError(t, err)
-	
-	_, err = writer.Write([]byte(content))
-	require.NoError(t, err)
-	
-	err = zipWriter.Close()
-	require.NoError(t, err)
-	
-	return &buf
-}
-
-func TestAPI_handleUpload_ZIPConvert(t *testing.T) {
-	coreTesting.RunTestCaseWithDB(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		internal.RegisterHashes()
-		
-		// Arrange
-		api := NewAPI()
-		require.NotNil(tb, api)
-		
-		uploadService := core.GetService[*mocks.MockUploadService](ctx, pluginCore.UPLOAD_SERVICE)
-		workflowService := core.GetService[*coreMocks.MockWorkflowService](ctx, core.WORKFLOW_SERVICE)
-		
-		// Create test ZIP file
-		zipBuf := createTestZIPFile(t, "Hello, World!")
-		
-		// Create multipart form
-		var body bytes.Buffer
-		writer := multipart.NewWriter(&body)
-		part, err := writer.CreateFormFile("file", "test.zip")
-		require.NoError(t, err)
-		
-		_, err = part.Write(zipBuf.Bytes())
-		require.NoError(t, err)
-		
-		err = writer.Close()
-		require.NoError(t, err)
-		
-		// Create HTTP request
-		req := httptest.NewRequest(http.MethodPost, "/api/upload?zip_mode=convert", &body)
-		req.Header.Set("Content-Type", writer.FormDataContentType())
-		
-		// Set up expectations
-		testCID := cid.MustParse("bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi")
-		uploadService.EXPECT().HandleUploadWithMode(mock.Anything, mock.Anything, mock.Anything, "convert").Return(testCID, "test-upload-id", nil)
-		workflowService.EXPECT().StartWorkflow(mock.Anything, mock.Anything, mock.Anything).Return(nil, nil)
-		
-		// Act
-		rec := httptest.NewRecorder()
-		c := echo.New().NewContext(req, rec)
-		
-		// Create a mock context for the API
-		mockCtx := &coreTesting.MockContext{}
-		mockCtx.On("DB").Return(ctx.DB())
-		mockCtx.On("Config").Return(ctx.Config())
-		mockCtx.On("APILogger", mock.Anything).Return(ctx.Logger())
-		
-		// Set up the API with the mock context
-		apiInstance, err := NewAPI()
-		require.NoError(t, err)
-		
-		// This is a simplified test - in a full implementation, we'd need to properly set up the echo context
-		// with authentication and other middleware
-		
-		// For now, we'll just test that the method exists and can be called
-		assert.NotNil(t, apiInstance)
-	})
-}
-
-func TestAPI_handleUpload_ZIPPreserve(t *testing.T) {
-	coreTesting.RunTestCaseWithDB(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		internal.RegisterHashes()
-		
-		// Arrange
-		api := NewAPI()
-		require.NotNil(tb, api)
-		
-		uploadService := core.GetService[*mocks.MockUploadService](ctx, pluginCore.UPLOAD_SERVICE)
-		workflowService := core.GetService[*coreMocks.MockWorkflowService](ctx, core.WORKFLOW_SERVICE)
-		
-		// Create test ZIP file
-		zipBuf := createTestZIPFile(t, "Hello, World!")
-		
-		// Create multipart form
-		var body bytes.Buffer
-		writer := multipart.NewWriter(&body)
-		part, err := writer.CreateFormFile("file", "test.zip")
-		require.NoError(t, err)
-		
-		_, err = part.Write(zipBuf.Bytes())
-		require.NoError(t, err)
-		
-		err = writer.Close()
-		require.NoError(t, err)
-		
-		// Create HTTP request
-		req := httptest.NewRequest(http.MethodPost, "/api/upload?zip_mode=preserve", &body)
-		req.Header.Set("Content-Type", writer.FormDataContentType())
-		
-		// Set up expectations
-		testCID := cid.MustParse("bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi")
-		uploadService.EXPECT().HandleUploadWithMode(mock.Anything, mock.Anything, mock.Anything, "preserve").Return(testCID, "test-upload-id", nil)
-		workflowService.EXPECT().StartWorkflow(mock.Anything, mock.Anything, mock.Anything).Return(nil, nil)
-		
-		// Act
-		rec := httptest.NewRecorder()
-		c := echo.New().NewContext(req, rec)
-		
-		// Create a mock context for the API
-		mockCtx := &coreTesting.MockContext{}
-		mockCtx.On("DB").Return(ctx.DB())
-		mockCtx.On("Config").Return(ctx.Config())
-		mockCtx.On("APILogger", mock.Anything).Return(ctx.Logger())
-		
-		// Set up the API with the mock context
-		apiInstance, err := NewAPI()
-		require.NoError(t, err)
-		
-		// This is a simplified test - in a full implementation, we'd need to properly set up the echo context
-		// with authentication and other middleware
-		
-		// For now, we'll just test that the method exists and can be called
-		assert.NotNil(t, apiInstance)
-	})
-}
-
-func TestAPI_handleUpload_InvalidZipMode(t *testing.T) {
-	coreTesting.RunTestCaseWithDB(t, func(tb coreTesting.TB, ctx coreTesting.TestContext) {
-		internal.RegisterHashes()
-		
-		// Arrange
-		api := NewAPI()
-		require.NotNil(tb, api)
-		
-		// Create test ZIP file
-		zipBuf := createTestZIPFile(t, "Hello, World!")
-		
-		// Create multipart form
-		var body bytes.Buffer
-		writer := multipart.NewWriter(&body)
-		part, err := writer.CreateFormFile("file", "test.zip")
-		require.NoError(t, err)
-		
-		_, err = part.Write(zipBuf.Bytes())
-		require.NoError(t, err)
-		
-		err = writer.Close()
-		require.NoError(t, err)
-		
-		// Create HTTP request with invalid zip_mode
-		req := httptest.NewRequest(http.MethodPost, "/api/upload?zip_mode=invalid", &body)
-		req.Header.Set("Content-Type", writer.FormDataContentType())
-		
-		// Act
-		rec := httptest.NewRecorder()
-		c := echo.New().NewContext(req, rec)
-		
-		// Create a mock context for the API
-		mockCtx := &coreTesting.MockContext{}
-		mockCtx.On("DB").Return(ctx.DB())
-		mockCtx.On("Config").Return(ctx.Config())
-		mockCtx.On("APILogger", mock.Anything).Return(ctx.Logger())
-		
-		// Set up the API with the mock context
-		apiInstance, err := NewAPI()
-		require.NoError(t, err)
-		
-		// This is a simplified test - in a full implementation, we'd need to properly set up the echo context
-		// with authentication and other middleware
-		
-		// For now, we'll just test that the method exists and can be called
-		assert.NotNil(t, apiInstance)
-	})
-}
+// TestOptions provides test configuration for API tests
+var TestOptions = coreTesting.CombineOptions(
+	coreTesting.WithMockServiceFactory(pluginCore.FILE_MANAGER_SERVICE, mocks.NewMockFileManagerService),
+	coreTesting.WithMockServiceFactory(pluginCore.PIN_SERVICE, mocks.NewMockIPFSPinService),
+	coreTesting.WithMockServiceFactory(pluginCore.BLOCK_SERVICE, mocks.NewMockBlockService),
+	coreTesting.WithMockServiceFactory(pluginCore.UPLOAD_SERVICE, mocks.NewMockUploadService),
+	coreTesting.WithMockServiceFactory(pluginCore.IPNS_KEY_SERVICE, mocks.NewMockIPNSKeyService),
+	coreTesting.WithMockServiceFactory(pluginCore.IPNS_PUBLISHER_SERVICE, mocks.NewMockIPNSPublisherService),
+	coreTesting.WithMockServiceFactory(pluginCore.WEBSITE_SERVICE, mocks.NewMockWebsiteService),
+	coreTesting.WithHTTPService(),
+	coreTesting.WithPlugins(),
+	coreTesting.WithAPIConfig(internal.ProtocolName, &pluginConfig.APIConfig{
+		GatewaySecret: "test-secret",
+	}),
+)
