+++ internal/api/dto/utils.go
@@ -3,9 +3,30 @@ package dto
 import (
 	"encoding/json"
 	"fmt"
+	"strings"
 
 	"github.com/multiformats/go-multiaddr"
 )
+
+const (
+	IPFSPathPrefix = "/ipfs/"
+	IPNSPathPrefix = "/ipns/"
+)
+
+// IPFSPath creates a properly formatted IPFS path from a CID string
+func IPFSPath(cid string) string {
+	return IPFSPathPrefix + trimPath(cid)
+}
+
+// IPNSPath creates a properly formatted IPNS path from a peer ID string
+func IPNSPath(peerID string) string {
+	return IPNSPathPrefix + trimPath(peerID)
+}
+
+// trimPath defensively trims leading and trailing slashes from path components
+func trimPath(s string) string {
+	return strings.Trim(s, "/")
+}
 
 // jsonToMap converts datatypes.JSON to map[string]string
 func jsonToMap(jsonData []byte) (map[string]string, error) {
