+++ internal/config/api.go
@@ -5,8 +5,12 @@ import "go.lumeweb.com/portal/config"
 var _ config.APIConfig = (*APIConfig)(nil)
 
 type APIConfig struct {
+	// GatewaySecret is the shared secret for authenticating gateway requests
+	GatewaySecret string `config:"gateway_secret"`
 }
 
 func (A APIConfig) Defaults() map[string]any {
-	return nil
+	return map[string]any{
+		"GatewaySecret": "",
+	}
 }
